Move FastAPI setup into app.py.
